web-api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
redis==5.2.1
python-dateutil==2.9.0.post0
loguru==0.7.3
yookassa==3.3.2
python-multipart==0.0.20
slowapi==0.1.9
